workflow-workers: address review feedback\n\n- Move graceful shutdown logic into helper.ts and register via registerGracefulShutdown(); escalate unhandledRejection to graceful shutdown; default timeout to 1 hour (configurable via SHUTDOWN_TIMEOUT_MS).\n- Update Dockerfile to use Corepack with pinned pnpm and multi-stage pnpm deploy; remove runtime pnpm and start with ; keep STOPSIGNAL SIGTERM.\n- docker-compose: set stop_grace_period to 1h and pass SHUTDOWN_TIMEOUT_MS=3600000 to align with long-running workflows.\n- Switch final CMD from pnpm to node to reduce runtime surface area.
